requirements.txt
pypdf
beautifulsoup4
google-genai
flask
scikit-learn
